+++ src/bundle/Resources/config/services/controllers.yaml
@@ -46,6 +46,4 @@ services:
 
     Ibexa\Bundle\User\Controller\UserInvitationController:
         autowire: true
-        calls:
-            - [ setContainer, [ '@service_container' ] ]
         tags: [ controller.service_arguments ]
