+++ src/lib/Form/Type/Invitation/SiteAccessChoiceType.php
@@ -6,27 +6,21 @@
  */
 declare(strict_types=1);
 
-namespace Ibexa\User\Form\Type;
+namespace Ibexa\User\Form\Type\Invitation;
 
-use Ibexa\Contracts\Core\Repository\Repository;
-use Ibexa\Contracts\Core\Repository\SectionService;
+use Ibexa\Core\MVC\Symfony\SiteAccess\SiteAccessServiceInterface;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\ChoiceList\ChoiceList;
 use Symfony\Component\Form\Extension\Core\Type\ChoiceType;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 
-final class SectionsChoiceType extends AbstractType
+final class SiteAccessChoiceType extends AbstractType
 {
-    private SectionService $sectionService;
+    private SiteAccessServiceInterface $siteAccessService;
 
-    private Repository $repository;
-
-    public function __construct(
-        Repository $repository,
-        SectionService $sectionService
-    ) {
-        $this->sectionService = $sectionService;
-        $this->repository = $repository;
+    public function __construct(SiteAccessServiceInterface $siteAccessService)
+    {
+        $this->siteAccessService = $siteAccessService;
     }
 
     public function configureOptions(OptionsResolver $resolver): void
@@ -35,10 +29,11 @@ final class SectionsChoiceType extends AbstractType
             ->setDefaults([
                 'choice_loader' => ChoiceList::lazy(
                     $this,
-                    fn () => $this->repository->sudo(fn () => $this->sectionService->loadSections())
+                    fn () => $this->siteAccessService->getAll(),
                 ),
                 'choice_label' => 'name',
-                'choice_value' => 'id',
+                'choice_name' => 'name',
+                'choice_value' => 'name',
             ]);
     }
